Refactor controller type to use strongly-typed enum
Replaces int/controller enum with uqm::ControllerType throughout the codebase for improved type safety and clarity. Updates CLI option to accept string values for controller type. Refactors UI and frame selection logic to use the new enum, including switch statements and helper functions. Disables resolution CLI option. Updates CLI argument parsing to use enum pairs. Modernizes controller handling and reduces risk of bugs from implicit integer usage.

+++ src/uqm/supermelee/buildpick.cpp
@@ -41,7 +41,7 @@ void BuildBuildPickFrame(void)
 	s.origin.x = 0;
 	s.origin.y = 0;
 	s.frame = SetAbsFrameIndex(MeleeFrame,
-							   BUILDPICK_FRAME_OFFSET + optControllerType);
+							   BUILDPICK_FRAME_OFFSET + static_cast<int>(optControllerType));
 	// 5x5 grid of ships to pick from
 	GetFrameRect(s.frame, &r);
 
@@ -115,7 +115,7 @@ void RepairBuildPickFrame(GFXRECT* pRect, GFXPOINT* origin)
 	s.origin.x = pRect->corner.x;
 	s.origin.y = pRect->corner.y;
 	s.frame = SetAbsFrameIndex(MeleeFrame,
-							   BUILDPICK_FRAME_OFFSET + optControllerType);
+							   BUILDPICK_FRAME_OFFSET + static_cast<int>(optControllerType));
 
 	DrawStamp(&s);
